Moved Transformations and Data types out and fixed shadow bug
