+++ src/Main.hs
@@ -14,10 +14,12 @@
 --
 module Main where
 
+import DataTypes
 import Vector
 import Quaternion
 import Color
 import Camera
+import Transformations
 import Utils (degreesToRadians, roots)
 import Codec.Picture
 import Codec.Picture.Png
@@ -25,39 +27,26 @@ import Data.Maybe
 import Data.List
 import Debug.Trace
 
--- Basic Data Types
-data Object = Object Shape Material deriving(Eq)
-data Material = Material Color deriving(Eq)
-data Shape = Sphere Vector Double  -- center, radius
-           | Plane Vector Vector  deriving(Eq) -- center, normal
-data Light = PointLight {center :: Vector, intensity :: Double} -- center, intensity
-data Scene = Scene [Object] [Light] Camera Config
-data Config = Config { sceneWidth :: Int,
-                       sceneHeight :: Int,
-                       defaultColor :: Color }
-
-data Ray = Ray {origin :: Vector, direction :: Vector}
-
 main :: IO()
 main =
   let
     objects :: [Object]
     objects = [Object
-                    (Sphere (Vector 0.5 0 2) 0.5)
+                    (Sphere (Vector (-0.5) 0 2) 0.5)
                     (Material Color.red),
-               Object
-                    (Sphere (Vector (-1) 0 4) 0.5)
-                    (Material Color.green),
                Object
                     (Sphere (Vector 0 0 3) 0.5)
                     (Material Color.blue),
                Object
                     (Plane (Vector 0 (-0.5) 0) (Vector 0 1 0))
+                    (Material Color.pink),
+               Object
+                    (Plane (Vector 0 0 5) (Vector 0 0 (-1)))
                     (Material Color.pink)
               ]
 
     lights :: [Light]
-    lights = [PointLight (Vector 0 0.5 0) 0.8, PointLight (Vector 0.5 0.5 0) 0.2]
+    lights = [PointLight (Vector 0 0.5 0) 0.4, PointLight (Vector 0.5 0.5 0) 0.4, PointLight (Vector 9 0 4) 0.2]
 
     camera :: Camera
     camera = Camera 45 (Vector 0 0 0) (Vector 0 0 1)
@@ -67,7 +56,7 @@ main =
     scene :: Scene
     scene = Scene objects lights camera config
 
-    img = generateImage (\x y -> pixelRGB8 $ Main.trace scene x (sceneHeight config - y)) (sceneWidth config) (sceneHeight config)
+    img = generateImage (\x y -> pixelRGB8 $ Main.trace scene (sceneWidth config - x) (sceneHeight config - y)) (sceneWidth config) (sceneHeight config)
    in
     writePng "output.png" img
 
@@ -80,26 +69,29 @@ trace (Scene objects lights camera config) x y =
     in
         case maybeIntersectedObject of
             Nothing -> backgroundColor
-            (Just intersectionObj@(_,object)) ->
+            Just intersectionObj  ->
                 getColorFromIntersection backgroundColor ray lights objects intersectionObj
 
-isLightVisible :: [Object] -> Vector -> Light -> Bool
-isLightVisible objects point light =
-    let
-        direction = normalize $ (center light) `sub` point
-        ray = Ray point direction
-        objIntersections = fmap (minIntersection ray) objects
-    in
-        all isNothing objIntersections
 
 getColorFromIntersection :: Color -> Ray -> [Light] -> [Object] -> (Double, Object) -> Color
-getColorFromIntersection defaultColor ray lights objects (hitDistance , hitObject) =
+getColorFromIntersection defaultColor ray lights objects (hitDistance, hitObject) =
     let
         hitPoint = pointAlongRay ray hitDistance
         otherObjects = filter (/= hitObject) objects
         visibleLights = filter (isLightVisible otherObjects hitPoint) lights
     in
         lambertColor hitPoint hitObject visibleLights
+
+isLightVisible :: [Object] -> Vector -> Light -> Bool
+isLightVisible objects point light =
+    let
+        toLightVector = center light `sub` point
+        distanceToLight = magnitude toLightVector
+        direction = normalize toLightVector
+        ray = Ray point direction
+        objIntersections = mapMaybe (minIntersection ray) objects
+    in
+        all ((\x -> x >= distanceToLight || x < 0) . fst) objIntersections
 
 -- Should also include light color
 lambertColor :: Vector -> Object -> [Light] -> Color
@@ -181,31 +173,3 @@ minIntersection ray@(Ray origin direction) object@(Object (Plane center normal) 
         distance = ((center `sub` origin) `dot` normal) / (direction `dot` normal)
         point = pointAlongRay ray distance
     in if distance < 0 then Nothing else Just (distance, object)
-
--- Transformations
-rotateObj :: Vector -> Double -> Object -> Object
-rotateObj axis angle (Object shape material) = Object (rotateShape axis angle shape) material
-
-rotateObjAroundPoint :: Vector -> Vector -> Double -> Object -> Object
-rotateObjAroundPoint point axis angle (Object shape material) = Object (rotateShapeAroundPoint point axis angle shape) material
-
-rotateShapeAroundPoint :: Vector -> Vector -> Double -> Shape -> Shape
-rotateShapeAroundPoint point axis angle shape = translateShape point (rotateShape axis angle (translateShape (neg point) shape))
-
-rotateShape :: Vector -> Double -> Shape -> Shape
-rotateShape axis angle (Sphere center radius) = Sphere (rotate axis angle center) radius
-rotateShape axis angle (Plane center normal) = Plane (rotate axis angle center) normal
-
-translateObj :: Vector -> Object -> Object
-translateObj translationVector (Object shape material) = Object (translateShape translationVector shape) material
-
-translateShape :: Vector -> Shape -> Shape
-translateShape translationVector (Sphere center radius) = Sphere (center `add` translationVector) radius
-translateShape translationVector (Plane center normal) = Plane (center `add` translationVector) normal
-
-scaleObj :: Object -> Double -> Object
-scaleObj (Object shape material) factor = Object (scaleShape factor shape) material
-
-scaleShape :: Double -> Shape -> Shape
-scaleShape factor (Sphere center radius) = Sphere center (factor * radius)
-scaleShape factor shape = shape
